Move debug/test hooks out of GhosttySurfaceView.swift for the length budget
The recovery-stress hooks and watchdog wiring pushed GhosttySurfaceView
past its file-length budget (4015 vs 3964). Relocate the stress snapshot
and force-recovery hooks to GhosttySurfaceView+RecoveryStress.swift and
the pre-existing testing helpers to GhosttySurfaceView+TestingHooks.swift
(the established +Extension pattern in this package), widening the
members they read to internal / private(set). The STUCK-log default
moves into SurfaceFreeDrainWatchdog itself, and the redundant
setFocusForRecoveryStress wrapper is deleted now that setFocus is
internal. No behavior change.
